requirements.txt
openai
hidapi
pyserial

